CORE/SCRIPTS: Fix for droprate, change type to uint16 and adjust relevan...

+++ src/map/utils/itemutils.h
@@ -45,7 +45,7 @@ struct DropItem_t
 {
 	uint16 ItemID;
 	uint8  DropType;
-	float  DropRate;
+	uint16  DropRate;
 };
 
 struct LootItem_t
